requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
channels==4.0.0
channels-redis==4.1.0
dj-database-url==2.0.0
Django==4.2.1
django-environ==0.10.0
django-extensions==3.2.1
django-htmx==1.14.0
djangorestframework==3.14.0
environ==1.0
et-xmlfile==1.1.0
gunicorn==20.1.0
msgpack==1.0.5
mysqlclient==2.1.1
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pip-review==1.3.0
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
whitenoise==6.4.0
